opengl/requirements.txt
numpy==1.21.2
Pillow==8.4.0
PyOpenGL==3.1.5
